requirements.txt
colorzero==2.0
cycler==0.11.0
fonttools==4.33.3
gpiozero==1.6.2
keyboard==0.13.5
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
picamera==1.13
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
